Setting Up JWT Handling (and so slight organization)
Implement the jwt handling in much the same way as Gio's diff. Also provides a plain text way to specify the content of the JWT via query param to make development easier. This is only available in development.

Hit `/api/v1/user?token={"id":"[someUserId]"}`
