+++ .env.example
@@ -3,3 +3,4 @@
 MONGODB_URI=
 NODE_ENV=
 DEBUG=
+JWT_SECRET={secret}
