+++ middleware/notFound.js
@@ -1,0 +1,5 @@
+const http = require('http');
+
+module.exports = (req, res) => {
+  res.status(404).json({ message: http.STATUS_CODES[404] });
+};
