+++ package.json
@@ -25,6 +25,7 @@
     "engine-strict": "1.0.0",
     "express": "4.16.3",
     "express-force-ssl": "0.3.2",
+    "express-jwt": "5.3.1",
     "express-restify-mongoose": "5.0.0",
     "feathers-mongoose": "6.1.2",
     "helmet": "3.12.1",
